Add the interactivity in a module
