+++ src/index.js
@@ -1,5 +1,9 @@
 /* eslint-disable arrow-body-style */
+/* eslint-disable import/no-cycle */
+/* eslint-disable import/no-mutable-exports */
+
 import './style.css';
+import { markCompleted, clearCompleted } from './modules/TaskCompletion.js';
 
 // Grabbing the DOM element for the task container
 const activities = document.querySelector('.activities');
@@ -10,11 +14,9 @@ const input = document.querySelector('.tsk');
 // Getting a form element
 const form = document.querySelector('#addForm');
 
-// getting clear button
-const clearBtn = document.querySelector('.clearBtn');
-
 // Declaring the array to hold the various tasks
-let tasks = JSON.parse(localStorage.getItem('data')) || [];
+/* eslint-disable import/prefer-default-export */
+export let tasks = JSON.parse(localStorage.getItem('data')) || [];
 
 // Add toDo to the local storage
 const addToDo = () => {
@@ -83,40 +85,6 @@ const edit = () => {
   });
 };
 
-// Check Completed value
-const markCompleted = () => {
-  let completedValue = false;
-  const checkInputs = document.querySelectorAll('.checkInput');
-  checkInputs.forEach((checkbox) => {
-    checkbox.addEventListener('click', () => {
-      /* eslint-disable no-unused-expressions */
-      checkbox.checked ? completedValue = true : completedValue = false;
-      const key = checkbox.parentElement.parentElement.getAttribute('key');
-      const result = tasks.find((task) => {
-        return String(key) === String(task.index);
-      });
-      // result.description = result.description;
-      result.completed = completedValue;
-      const convertedTasks = JSON.stringify(tasks);
-      localStorage.setItem('data', convertedTasks);
-      window.location.reload();
-    });
-  });
-};
-
-// Clear All Completed Tasks
-clearBtn.addEventListener('click', () => {
-  const uncompletedTasks = tasks.filter((item) => item.completed === false);
-
-  tasks = uncompletedTasks;
-  if (uncompletedTasks.length) {
-    localStorage.setItem('data', JSON.stringify(uncompletedTasks));
-  } else {
-    localStorage.clear();
-  }
-  window.location.reload();
-});
-
 // Remove Section
 const remove = () => {
   const deleteButtons = document.querySelectorAll('.delete');
@@ -147,4 +115,5 @@ addToDo();
 displayToDos();
 edit();
 markCompleted();
+clearCompleted();
 remove();
